mabin/backend/requirements.txt
Django
djangorestframework
django-cors-headers
djangorestframework-simplejwt
PyJWT
cryptography
Pillow
python-magic
python-dateutil
pytz
django-filter
drf-yasg
django-debug-toolbar
openpyxl
